requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
torch==2.0.1
ultralytics==8.0.145
pytesseract==0.3.10
whisper==1.0.0
sentence-transformers==2.2.2
psycopg2-binary==2.9.6
pytest==7.3.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
logging==0.5.1.2